# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/user/login": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "Json请求体",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "Json请求体",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.UserBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/user/searchFriends": {
            "get": {
                "tags": [
                    "用户模块"
                ],
                "summary": "查询联系人",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.UserBody": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "utils.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据结构体"
                },
                "msg": {
                    "description": "响应说明",
                    "type": "string"
                }
            }
        }
    }
}